move files to debug.py
